requirements.txt
# R&D Tax Agent - Python Dependencies

# GitHub Integration
PyGithub==2.1.1
GitPython==3.1.40

# Azure OpenAI
openai==1.3.0

# Vector Database
chromadb==0.4.18

# Document Generation
python-docx==1.1.0

# Optional: Enhanced document templating
# docxtpl==0.16.7
# Jinja2==3.1.2

# Utilities
python-dateutil==2.8.2
typing-extensions==4.8.0
